+++ docs/guides/debugging.md
@@ -6,9 +6,12 @@
   * [API Overview](#api-overview)
   * [Log Safely](#log-safely)
   * [Log Objects Usefully](#log-objects-usefully)
+  * [Creating new Loggers](#creating-new-loggers)
   * [Log Levels](#log-levels)
   * [Available Log Levels](#available-log-levels)
+  * [Debug Logging](#debug-logging)
   * [History](#history)
+    * [History filtering](#history-filtering)
 
 ## Logging
 
@@ -18,17 +21,19 @@ Video.js includes `videojs.log`, a lightweight wrapper around a subset of [the `
 
 Most of these methods should be fairly self-explanatory, but for complete details, see [the API docs][api].
 
-| Method                          | Alias Of         | Matching Level(s)              |
-| ------------------------------- | ---------------- | ------------------------------ |
-| `videojs.log()`                 | `console.log`    | all, debug, info               |
-| `videojs.log.debug()`           | `console.debug`  | all, debug                     |
-| `videojs.log.warn()`            | `console.warn`   | all, debug, info, warn         |
-| `videojs.log.error()`           | `console.error`  | all, debug, info, warn, error  |
-| `videojs.log.level()`           | n/a              | n/a                            |
-| `videojs.log.history()`         | n/a              | n/a                            |
-| `videojs.log.history.clear()`   | n/a              | n/a                            |
-| `videojs.log.history.disable()` | n/a              | n/a                            |
-| `videojs.log.history.enable()`  | n/a              | n/a                            |
+| Method                          | Alias Of        | Matching Level(s)             |
+| ------------------------------- | --------------- | ----------------------------- |
+| `videojs.log()`                 | `console.log`   | all, debug, info              |
+| `videojs.log.debug()`           | `console.debug` | all, debug                    |
+| `videojs.log.warn()`            | `console.warn`  | all, debug, info, warn        |
+| `videojs.log.error()`           | `console.error` | all, debug, info, warn, error |
+| `videojs.log.createLogger()`    | n/a             | n/a                           |
+| `videojs.log.level()`           | n/a             | n/a                           |
+| `videojs.log.history()`         | n/a             | n/a                           |
+| `videojs.log.history.clear()`   | n/a             | n/a                           |
+| `videojs.log.history.disable()` | n/a             | n/a                           |
+| `videojs.log.history.enable()`  | n/a             | n/a                           |
+| `videojs.log.history.filter()`  | n/a             | n/a                           |
 
 For descriptions of these features, please refer to the sections below.
 
@@ -54,6 +59,23 @@ it will produce this:
 
 ```txt
 VIDEOJS: this is a string {"butThis": "is an object"}
+```
+
+### Creating new Loggers
+
+Sometimes, you want to make a new module or plugin and log messages with a label. Kind of how all these logs are prepended with `VIDEOJS:`. You can do that via the `createLogger` method. It takes a name and gives you back a log object like `videojs.log`. Here's an example:
+
+```js
+const mylogger = videojs.log.createLogger('mylogger');
+
+mylogger('hello world!');
+// > VIDEOJS: mylogger: hello world!
+
+// We can even chain it further
+const anotherlogger = mylogger.createLogger('anotherlogger');
+
+anotherlogger('well, hello there');
+// > VIDEOJS: mylogger: anotherlogger: well, hello there
 ```
 
 ### Log Levels
@@ -90,7 +112,7 @@ Although the log levels attempt to match their `window.console` counterparts, `w
 
 ### History
 
-> **Note:** In Video.js 5, `videojs.log.history` was an array. As of Video.js 6, it is a function which returns an array. This change was made to provide a richer, safer logging history API.
+> **Note:** In Video.js 5, `videojs.log.history` was an array. As of Video.js 6, it is a function which returns an array. This change was made to provide a richer, safer logging history API. You can also filter the history based on the name of the logger.
 
 By default, the `videojs.log` module tracks a history of _everything_ passed to it regardless of logging level:
 
@@ -112,6 +134,28 @@ Finally, the history (if enabled) can be cleared at any time via:
 
 ```js
 videojs.log.history.clear();
+```
+
+#### History filtering
+
+If you want to find all the history that was created by a particular logger, you can do so via `history.filter()`.
+Given a specific logger with name `foo`, you can pass `foo` to `history.filter()` and get all items logger by foo.
+Let me show you an example:
+
+```js
+const mylogger = videojs.log.createLogger('mylogger');
+const anotherlogger = mylogger.createLogger('anotherlogger');
+
+videojs.log('hello');
+mylogger('how are you');
+anotherlogger('today');
+
+videojs.log.history.filter('VIDEOJS');
+// > [['VIDEOJS:', 'hello'], ['VIDEOJS: mylogger:', 'how are you'], ['VIDEOJS: mylogger: anotherlogger:', 'today']]
+videojs.log.history.filter('mylogger');
+// > [['VIDEOJS:    mylogger:', 'how are you'], ['VIDEOJS: mylogger: anotherlogger:', 'today']]
+videojs.log.history.filter('anotherlogger');
+// > [['VIDEOJS: mylogger: anotherlogger:', 'today']]
 ```
 
 [api]: https://docs.videojs.com/
