+++ test/unit/utils/log.test.js
@@ -1,7 +1,6 @@
 /* eslint-env qunit */
 import {IE_VERSION} from '../../../src/js/utils/browser';
 import log from '../../../src/js/utils/log.js';
-import {logByType} from '../../../src/js/utils/log.js';
 import window from 'global/window';
 import sinon from 'sinon';
 
@@ -65,30 +64,30 @@ QUnit.test('logging functions should work', function(assert) {
   assert.ok(window.console.warn.called, 'warn was called');
   assert.deepEqual(
     window.console.warn.firstCall.args,
-    getConsoleArgs('VIDEOJS:', 'WARN:', 'warn1', 'warn2')
+    ['VIDEOJS:', 'WARN:', 'warn1', 'warn2']
   );
 
   assert.ok(window.console.error.called, 'error was called');
   assert.deepEqual(
     window.console.error.firstCall.args,
-    getConsoleArgs('VIDEOJS:', 'ERROR:', 'error1', 'error2')
+    ['VIDEOJS:', 'ERROR:', 'error1', 'error2']
   );
 
   const history = log.history();
 
   assert.equal(history.length, 4, 'there should be four messages in the log history');
   assert.deepEqual(history[0],
-                   ['log1', 'log2'],
+                   ['VIDEOJS:', 'log1', 'log2'],
                    'history recorded the correct arguments');
   // although not enabled by default, history should still maintain the record
   assert.deepEqual(history[1],
-                   ['DEBUG:', 'debug1', 'debug2'],
+                   ['VIDEOJS:', 'DEBUG:', 'debug1', 'debug2'],
                    'history recorded the correct arguments');
   assert.deepEqual(history[2],
-                   ['WARN:', 'warn1', 'warn2'],
+                   ['VIDEOJS:', 'WARN:', 'warn1', 'warn2'],
                    'history recorded the correct arguments');
   assert.deepEqual(history[3],
-                   ['ERROR:', 'error1', 'error2'],
+                   ['VIDEOJS:', 'ERROR:', 'error1', 'error2'],
                    'history recorded the correct arguments');
 });
 
@@ -99,18 +98,21 @@ QUnit.test('in IE pre-11 (or when requested) objects and arrays are stringified'
   log.history.clear();
 
   // Run a custom log call, explicitly requesting object/array stringification.
-  logByType('log', [
+  log.stringify = true;
+  log(
     'test',
     {foo: 'bar'},
     [1, 2, 3],
     0,
     false,
     null
-  ], true);
+  );
 
   assert.ok(window.console.log.called, 'log was called');
   assert.deepEqual(window.console.log.firstCall.args,
             ['VIDEOJS: test {"foo":"bar"} [1,2,3] 0 false null']);
+
+  delete log.stringify;
 });
 
 QUnit.test('setting the log level changes what is actually logged', function(assert) {
@@ -131,9 +133,9 @@ QUnit.test('setting the log level changes what is actually logged', function(ass
 
   const history = log.history();
 
-  assert.deepEqual(history[0], ['log1', 'log2'], 'history is maintained even when logging is not performed');
-  assert.deepEqual(history[1], ['WARN:', 'warn1', 'warn2'], 'history is maintained even when logging is not performed');
-  assert.deepEqual(history[2], ['ERROR:', 'error1', 'error2'], 'history is maintained even when logging is not performed');
+  assert.deepEqual(history[0], ['VIDEOJS:', 'log1', 'log2'], 'history is maintained even when logging is not performed');
+  assert.deepEqual(history[1], ['VIDEOJS:', 'WARN:', 'warn1', 'warn2'], 'history is maintained even when logging is not performed');
+  assert.deepEqual(history[2], ['VIDEOJS:', 'ERROR:', 'error1', 'error2'], 'history is maintained even when logging is not performed');
 
   log.level('off');
 
@@ -197,10 +199,10 @@ QUnit.test('supports debug logging', function(assert) {
   const history = log.history();
 
   assert.equal(history.length, 4, 'four messages in history');
-  assert.deepEqual(history[0], ['log1', 'log2'], 'history is maintained');
-  assert.deepEqual(history[1], ['DEBUG:', 'debug1', 'debug2'], 'history is maintained');
-  assert.deepEqual(history[2], ['WARN:', 'warn1', 'warn2'], 'history is maintained');
-  assert.deepEqual(history[3], ['ERROR:', 'error1', 'error2'], 'history is maintained');
+  assert.deepEqual(history[0], ['VIDEOJS:', 'log1', 'log2'], 'history is maintained');
+  assert.deepEqual(history[1], ['VIDEOJS:', 'DEBUG:', 'debug1', 'debug2'], 'history is maintained');
+  assert.deepEqual(history[2], ['VIDEOJS:', 'WARN:', 'warn1', 'warn2'], 'history is maintained');
+  assert.deepEqual(history[3], ['VIDEOJS:', 'ERROR:', 'error1', 'error2'], 'history is maintained');
 });
 
 QUnit.test('falls back to info and log when debug is not supported', function(assert) {
@@ -211,29 +213,31 @@ QUnit.test('falls back to info and log when debug is not supported', function(as
   log.level('debug');
 
   window.console.debug = null;
-  logByType('debug', ['debug1', 'debug2']);
+  log.debug('debug1', 'debug2');
 
   assert.ok(window.console.info.called, 'info was called');
   assert.notOk(window.console.log.called, 'log was not called');
   assert.notOk(window.console.warn.called, 'warn was not called');
   assert.notOk(window.console.error.called, 'error was not called');
-  assert.deepEqual(window.console.info.firstCall.args,
-                   getConsoleArgs('VIDEOJS:', 'DEBUG:', 'debug1', 'debug2'),
-                   'logged the right message');
+  assert.deepEqual(
+    window.console.info.firstCall.args,
+    ['VIDEOJS:', 'DEBUG:', 'debug1', 'debug2'],
+    'logged the right message'
+  );
 
   window.console.info = null;
-  logByType('debug', ['debug3', 'debug4']);
+  log.debug('debug3', 'debug4');
 
   assert.ok(window.console.log.called, 'log was called');
   assert.notOk(window.console.warn.called, 'warn was not called');
   assert.notOk(window.console.error.called, 'error was not called');
   assert.deepEqual(window.console.log.firstCall.args,
-                   getConsoleArgs('VIDEOJS:', 'DEBUG:', 'debug3', 'debug4'),
+                   ['VIDEOJS:', 'DEBUG:', 'debug3', 'debug4'],
                    'logged the right message');
 
   // when no comparable level logs are available, there should not be any logging
   window.console.log = null;
-  logByType('debug', ['debug5', 'debug6']);
+  log.debug('debug5', 'debug6');
 
   assert.notOk(window.console.warn.called, 'warn was not called');
   assert.notOk(window.console.error.called, 'error was not called');
